Improve r.js build times
It seems that the r.js optimiser attempts to parse the entire JS file for any dependencies, which makes the build very slow (>50 seconds!) given that the large ICAO airports and waypoints databases were previously defined as JS files.  We now instead use the JSON plugin to load JSON files and then only put it as a module in the optimised file.  Also update the text plugin and use Almond in the built file instead of RequireJS.